autonomous_data_analyst/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pandas==2.1.4
polars==0.20.6
numpy==1.25.2
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.2.0
plotly==5.17.0
vega-lite==5.14.0
openai==1.6.1
anthropic==0.8.1
transformers==4.36.2
torch==2.1.2
prophet==1.1.5
faiss-cpu==1.7.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0
